--- LIvecoding-uwp/LIvecoding-uwp/Models/MenuItem.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Windows.UI.Xaml.Controls;
-
-namespace LIvecoding_uwp.Models
-{
-    public class MenuItem
-    {
-        public Symbol Icon { get; set; }
-        public string Name { get; set; }
-        public Type PageType { get; set; }
-    }
-}
