--- LIvecoding-uwp/LIvecoding-uwp/Views/LiveStreamPage.xaml
@@ -1,64 +1,0 @@
-﻿<Page
-    x:Class="LIvecoding_uwp.Views.LiveStreamPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:LIvecoding_uwp.Views"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    DataContext="{Binding _liveStreamViewModel, Source={StaticResource Locator}}">
-    <Page.Resources>
-        <DataTemplate x:Key="DefaultTemplate">
-            <Grid Background="White" BorderBrush="#FF999595" BorderThickness="1">
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="5*"/>
-                    <RowDefinition/>
-                    <RowDefinition/>
-                </Grid.RowDefinitions>
-                <Grid Background="Black">
-                    <Grid.RowDefinitions>
-                        <RowDefinition/>
-                        <RowDefinition Height="1.5*"/>
-                    </Grid.RowDefinitions>
-                    <StackPanel Orientation="Horizontal" Width="50" Height="24" VerticalAlignment="Top" Background="Red" HorizontalAlignment="Left">
-                        <Border Height="10" Width="10" Margin="5,0,2,0" CornerRadius="15" Background="White" VerticalAlignment="Center"/>
-                        <TextBlock Text="LIVE" Foreground="White" FontSize="15" FontWeight="Bold" VerticalAlignment="Top"/>
-                    </StackPanel>
-                    <StackPanel Grid.Row="1" VerticalAlignment="Top" HorizontalAlignment="Center">
-                        <Border Height="90" Width="90" BorderBrush="Gray" BorderThickness="1" CornerRadius="50">
-                            <Border.Background>
-                                <ImageBrush Stretch="Fill" ImageSource="ms-appx:///Assets/userdashboard-img.png"/>
-                            </Border.Background>
-                        </Border>
-                        <TextBlock Text="{Binding Path=UserUrl}" FontSize="25" Foreground="White" FontWeight="Bold"/>
-                    </StackPanel>
-                </Grid>
-                <TextBlock Grid.Row="1" Text="{Binding Path=Title}" VerticalAlignment="Center" Foreground="Blue" FontSize="15" FontWeight="ExtraBold" Margin="10,0,0,0"/>
-                <Grid Grid.Row="2" Margin="10,0,10,0">
-                    <Grid.ColumnDefinitions>
-                        <ColumnDefinition/>
-                        <ColumnDefinition/>
-                    </Grid.ColumnDefinitions>
-                    <TextBlock Text="{Binding Path=Language}" Foreground="Gray" VerticalAlignment="Top"/>
-                    <StackPanel Grid.Column="1" Orientation="Horizontal" VerticalAlignment="Top" HorizontalAlignment="Right">
-                        <SymbolIcon Symbol="View"  VerticalAlignment="Top"/>
-                        <TextBlock Text="{Binding Path=ViewersLive}"  Foreground="Gray" VerticalAlignment="Bottom"/>
-                    </StackPanel>
-                </Grid>
-            </Grid>
-        </DataTemplate>
-    </Page.Resources>
-
-    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-        <ListView ItemTemplate="{StaticResource DefaultTemplate}" 
-                  ItemsSource="{Binding _videos}">
-            <ListView.ItemContainerStyle>
-                <Style TargetType="ListViewItem">
-                    <Setter Property="HorizontalContentAlignment" Value="Stretch"/>
-                    <Setter Property="Margin" Value="0,5,0,5"/>
-                </Style>
-            </ListView.ItemContainerStyle>
-
-        </ListView>
-    </Grid>
-</Page>
