+++ LIvecoding-uwp/LIvecoding-uwp/Views/LoginPage.xaml.cs
@@ -1,7 +1,4 @@
-﻿using LIvecoding_uwp.Configuration;
-using LivecodingApi.Model;
-using LivecodingApi.Services;
-using System;
+﻿using System;
 using Windows.UI.Xaml.Controls;
 
 // Pour plus d'informations sur le modèle d'élément Page vierge, voir la page http://go.microsoft.com/fwlink/?LinkId=234238
@@ -16,32 +13,6 @@ namespace LIvecoding_uwp.Views
         public LoginPage()
         {
             this.InitializeComponent();
-            ((ViewModels.LoginViewModel)DataContext).ChargerMenu();
-            contentFrame.Navigate(typeof(LiveStreamPage));
-            //Authenticate();
-        }
-
-        public void Authenticate()
-
-        {
-            var service = new ReactiveLivecodingApiService();
-            service.Login(AuthConstants.ClientId, AuthConstants.ClientSecret, new[] { AuthenticationScope.Read })
-
-                .Subscribe((result) =>
-
-                {
-                    
-                },
-
-                (error) =>
-
-                {
-
-                    throw new Exception();
-
-                });
-            contentFrame.Navigate(typeof(LiveStreamPage)); 
-
         }
     }
 }
