+++ LIvecoding-uwp/LIvecoding-uwp/Views/MainPage.xaml
@@ -1,13 +1,11 @@
 ﻿<Page
-    x:Class="LIvecoding_uwp.Views.LoginPage"
+    x:Class="LIvecoding_uwp.Views.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:LIvecoding_uwp.Views"
-    xmlns:controls="using:Microsoft.Toolkit.Uwp.UI.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    DataContext="{Binding Login, Source={StaticResource Locator}}">
+    mc:Ignorable="d">
 
     <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
