+++ LIvecoding-uwp/LIvecoding-uwp/project.json
@@ -2,7 +2,7 @@
   "dependencies": {
     "livecoding-api-pcl": "1.1.8",
     "Microsoft.NETCore.UniversalWindowsPlatform": "5.2.2",
-    "Microsoft.Toolkit.Uwp.UI.Controls": "1.0.0",
+    "Microsoft.Toolkit.Uwp.UI.Controls": "1.2.0",
     "MvvmLightLibs": "5.3.0"
   },
   "frameworks": {
